image: rework string serialization
Move the image string serialization closer to the approach used
by the plot hierarchy. This does mean that the output changes,
in that there is no longer a blank line added at the end.

This is the minimal change needed to take advantage of the
same logic to "convert a NumPy array to a string", rather than have
it repeated (although it is not a lot of code). Further combining
code could be done but it does not feel worth it here.
